Create ExecutorIntegrationTestBase, separate bedrock executor tests and update test configurations
