+++ .github/workflows/heavy-tests.yml
@@ -29,11 +29,11 @@ jobs:
       matrix:
         include:
           - job-name: "agent-tests"
-            test-group: "ai.koog.integration.tests.agent.* --exclude-tests '*BedrockNova*'"
+            test-group: "ai.koog.integration.tests.agent.* --exclude-tests ai.koog.integration.tests.agent.BedrockNovaAgentIntegrationTest"
             artifact-name: "agent-tests"
             os: ubuntu-latest
           - job-name: "single-llm-executor-tests"
-            test-group: "ai.koog.integration.tests.executor.SingleLLMPromptExecutorIntegrationTest --exclude-tests '*BedrockExecutor*'"
+            test-group: "ai.koog.integration.tests.executor.SingleLLMPromptExecutorIntegrationTest"
             artifact-name: "single-llm-executor-tests"
             os: ubuntu-latest
           - job-name: "multiple-llm-executor-tests"
@@ -41,11 +41,11 @@ jobs:
             artifact-name: "multiple-llm-executor-tests"
             os: ubuntu-latest
           - job-name: "bedrock-tests"
-            test-group: "*BedrockCombinations* *BedrockExecutor* *BedrockNova*"
+            test-group: "ai.koog.integration.tests.executor.BedrockExecutorIntegrationTest ai.koog.integration.tests.agent.BedrockNovaAgentIntegrationTest"
             artifact-name: "bedrock-tests"
             os: ubuntu-latest
           - job-name: "other-executor-tests"
-            test-group: "ai.koog.integration.tests.executor.* --exclude-tests ai.koog.integration.tests.executor.SingleLLMPromptExecutorIntegrationTest --exclude-tests ai.koog.integration.tests.executor.MultipleLLMPromptExecutorIntegrationTest --exclude-tests ai.koog.integration.tests.executor.OllamaExecutorIntegrationTest"
+            test-group: "ai.koog.integration.tests.executor.* --exclude-tests ai.koog.integration.tests.executor.SingleLLMPromptExecutorIntegrationTest --exclude-tests ai.koog.integration.tests.executor.MultipleLLMPromptExecutorIntegrationTest --exclude-tests ai.koog.integration.tests.executor.BedrockExecutorIntegrationTest --exclude-tests ai.koog.integration.tests.executor.OllamaExecutorIntegrationTest"
             artifact-name: "other-executor-tests"
             os: ubuntu-latest
           - job-name: "capabilities-tests"
